# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication is handled using JWT tokens. The tokens are set as access-token and refresh-token cookies.\nFor information regarding which role is authorized for which endpoint, see the http handler https://github.com/Zibbp/ganymede/blob/main/internal/transport/http/handler.go.",
        "title": "Ganymede API",
        "contact": {
            "name": "Zibbp",
            "url": "https://github.com/zibbp/ganymede"
        },
        "license": {
            "name": "GPL-3.0"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/queue": {
            "get": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Get queue items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Get queue items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get processing queue items",
                        "name": "processing",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Queue"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Create a queue item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Create a queue item",
                "parameters": [
                    {
                        "description": "Create queue item",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateQueueRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ent.Queue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/queue/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Get queue item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Get queue item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Queue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Update queue item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Update queue item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update queue item",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateQueueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Queue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Delete queue item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Delete queue item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/queue/{id}/tail": {
            "get": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Read queue log file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queue"
                ],
                "summary": "Read queue log file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Queue item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Log type: video, video-convert, chat, chat-render, or chat-convert",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/task/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyCookieAuth": []
                    }
                ],
                "description": "Start a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Start a task",
                "parameters": [
                    {
                        "description": "StartTaskRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.StartTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Queue": {
            "type": "object",
            "properties": {
                "chat_processing": {
                    "description": "ChatProcessing holds the value of the \"chat_processing\" field.",
                    "type": "boolean"
                },
                "chat_start": {
                    "description": "ChatStart holds the value of the \"chat_start\" field.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the QueueQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.QueueEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "string"
                },
                "live_archive": {
                    "description": "LiveArchive holds the value of the \"live_archive\" field.",
                    "type": "boolean"
                },
                "on_hold": {
                    "description": "OnHold holds the value of the \"on_hold\" field.",
                    "type": "boolean"
                },
                "processing": {
                    "description": "Processing holds the value of the \"processing\" field.",
                    "type": "boolean"
                },
                "render_chat": {
                    "description": "RenderChat holds the value of the \"render_chat\" field.",
                    "type": "boolean"
                },
                "task_chat_convert": {
                    "description": "TaskChatConvert holds the value of the \"task_chat_convert\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_chat_download": {
                    "description": "TaskChatDownload holds the value of the \"task_chat_download\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_chat_move": {
                    "description": "TaskChatMove holds the value of the \"task_chat_move\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_chat_render": {
                    "description": "TaskChatRender holds the value of the \"task_chat_render\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_video_convert": {
                    "description": "TaskVideoConvert holds the value of the \"task_video_convert\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_video_download": {
                    "description": "TaskVideoDownload holds the value of the \"task_video_download\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_video_move": {
                    "description": "TaskVideoMove holds the value of the \"task_video_move\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_vod_create_folder": {
                    "description": "TaskVodCreateFolder holds the value of the \"task_vod_create_folder\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_vod_download_thumbnail": {
                    "description": "TaskVodDownloadThumbnail holds the value of the \"task_vod_download_thumbnail\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_vod_save_info": {
                    "description": "TaskVodSaveInfo holds the value of the \"task_vod_save_info\" field.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt holds the value of the \"updated_at\" field.",
                    "type": "string"
                },
                "video_processing": {
                    "description": "VideoProcessing holds the value of the \"video_processing\" field.",
                    "type": "boolean"
                }
            }
        },
        "ent.QueueEdges": {
            "type": "object",
            "properties": {
                "vod": {
                    "description": "Vod holds the value of the vod edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Vod"
                        }
                    ]
                }
            }
        },
...
        "http.CreateQueueRequest": {
            "type": "object",
            "required": [
                "vod_id"
            ],
            "properties": {
                "vod_id": {
                    "type": "string"
                }
            }
        },
...
        "http.StartTaskRequest": {
            "type": "object",
            "required": [
                "task"
            ],
            "properties": {
                "task": {
                    "type": "string",
                    "enum": [
                        "check_live",
                        "check_vod",
                        "get_jwks",
                        "twitch_auth",
                        "queue_hold_check",
                        "storage_migration"
                    ]
                }
            }
        },
...
        "http.UpdateQueueRequest": {
            "type": "object",
            "required": [
                "task_chat_convert",
                "task_chat_download",
                "task_chat_move",
                "task_chat_render",
                "task_video_convert",
                "task_video_download",
                "task_video_move",
                "task_vod_create_folder",
                "task_vod_download_thumbnail",
                "task_vod_save_info"
            ],
            "properties": {
                "chat_processing": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "live_archive": {
                    "type": "boolean"
                },
                "on_hold": {
                    "type": "boolean"
                },
                "processing": {
                    "type": "boolean"
                },
                "task_chat_convert": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_chat_download": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_chat_move": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_chat_render": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_video_convert": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_video_download": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_video_move": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_vod_create_folder": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_vod_download_thumbnail": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "task_vod_save_info": {
                    "enum": [
                        "pending",
                        "running",
                        "success",
                        "failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.TaskStatus"
                        }
                    ]
                },
                "video_processing": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "utils.TaskStatus": {
            "type": "string",
            "enum": [
                "success",
                "running",
                "pending",
                "failed"
            ],
            "x-enum-varnames": [
                "Success",
                "Running",
                "Pending",
                "Failed"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyCookieAuth": {
            "type": "apiKey",
            "name": "access-token",
            "in": "cookie"
        }
    }
}